+++ releasenotes.md
@@ -1,12 +1,25 @@
-# ErrPref Release Notes Version 0.0.0
+# ErrPref Release Notes Version 1.1.0
+
+## Version 1.1.0
 
 This version of ***ErrPref*** was compiled and tested using Go 1.15.8.
 
-Not much to see here folks. This is the first release of the 'errpref'
-package. 
+### ErrPrefixDto
 
-All tests are passing, and beginning tomorrow, I will be using the types
-provided here in a real-world environment. 
+1. Converted method ErrPrefixDto.String() from a pointer receiver to a value receiver.
 
-Stay tuned ...
+2. Added 'X' Methods:
+   - XCtx() - Returns pointer to current instance.
 
+   - XEPref() - Returns pointer to current instance.
+
+   - XEPrefCtx() - Returns pointer to current instance.
+
+   - XEPrefOld() - Returns pointer to current instance.
+
+     
+
+
+
+## Version 0.0.0 
+Initial release of ***ErrPref***.